// src/interface/formTypes.d.ts
export interface Pasien {
    nik: string
    nama: string
    tglLahir: string
    tmptLahir: string
    jKelamin: string
    alamat: string
    pekerjaan: string
    nikah: string
    agama: string
    golDarah: string
    rt: string
    rw: string
    provinsi: string
    kelurahan: string
    kecamatan: string
    pendidikan: string
    noAsuransi: string
    ibuKandung: string
}